requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
tensorflow>=2.12.0
transformers>=4.30.0
pandas>=2.0.0
scikit-learn>=1.2.2
nltk>=3.8.1
torch>=2.0.0
numpy>=1.24.3
python-dotenv>=1.0.0
psycopg2>=2.9.10
psycopg2-binary>=2.9.10